+++ demo/admin/src/pages/EditPage.tsx
@@ -127,8 +127,7 @@ export const EditPage: React.FC<Props> = ({ id, category }) => {
     }
 
     return (
-<<<<<<< HEAD
-        <>
+        <AzureAiTranslatorProvider showApplyTranslationDialog={true} enabled={true}>
             {hasChanges && (
                 <RouterPrompt
                     message={(location) => {
@@ -203,86 +202,6 @@ export const EditPage: React.FC<Props> = ({ id, category }) => {
                 </BlockPreviewWithTabs>
             </MainContent>
             {dialogs}
-        </>
-=======
-        <AzureAiTranslatorProvider showApplyTranslationDialog={true} enabled={true}>
-            <EditPageLayout>
-                {hasChanges && (
-                    <RouterPrompt
-                        message={(location) => {
-                            if (location.pathname.startsWith(match.url)) return true; //we navigated within our self
-                            return intl.formatMessage(messages.saveUnsavedChanges);
-                        }}
-                        saveAction={handleSaveAction}
-                    />
-                )}
-                <Toolbar>
-                    <ToolbarItem>
-                        <IconButton onClick={stackApi?.goBack}>
-                            <ArrowLeft />
-                        </IconButton>
-                    </ToolbarItem>
-                    <PageName pageId={id} />
-                    <ToolbarFillSpace />
-                    <ToolbarActions>
-                        <Button
-                            color="info"
-                            startIcon={<Preview />}
-                            disabled={!pageState}
-                            onClick={() => {
-                                openSitePreviewWindow(pageState.path, contentScopeMatch.url);
-                            }}
-                        >
-                            <FormattedMessage id="pages.pages.page.edit.preview" defaultMessage="Web preview" />
-                        </Button>
-                        {pageSaveButton}
-                    </ToolbarActions>
-                </Toolbar>
-                <MainContent disablePaddingBottom>
-                    <BlockPreviewWithTabs previewUrl={`${siteConfig.previewUrl}/admin/page`} previewState={previewState} previewApi={previewApi}>
-                        {[
-                            {
-                                key: "content",
-                                label: (
-                                    <AdminTabLabel isValid={rootBlocksApi.content.isValid}>
-                                        <FormattedMessage {...messages.content} />
-                                    </AdminTabLabel>
-                                ),
-                                content: (
-                                    <AdminComponentRoot title={intl.formatMessage(messages.page)}>{rootBlocksApi.content.adminUI}</AdminComponentRoot>
-                                ),
-                            },
-                            {
-                                key: "config",
-                                label: (
-                                    <AdminTabLabel isValid={rootBlocksApi.seo.isValid}>
-                                        <FormattedMessage id="pages.pages.page.edit.config" defaultMessage="Config" />
-                                    </AdminTabLabel>
-                                ),
-                                content: rootBlocksApi.seo.adminUI,
-                            },
-                            {
-                                key: "dependencies",
-                                label: (
-                                    <AdminTabLabel isValid={rootBlocksApi.seo.isValid}>
-                                        <FormattedMessage id="pages.pages.page.edit.dependencies" defaultMessage="Dependencies" />
-                                    </AdminTabLabel>
-                                ),
-                                content: (
-                                    <DependencyList
-                                        query={pageDependenciesQuery}
-                                        variables={{
-                                            id: pageState?.document?.id ?? "",
-                                        }}
-                                    />
-                                ),
-                            },
-                        ]}
-                    </BlockPreviewWithTabs>
-                </MainContent>
-                {dialogs}
-            </EditPageLayout>
         </AzureAiTranslatorProvider>
->>>>>>> main
     );
 };