+++ packages/admin/admin/src/table/TableLocalChanges.tsx
@@ -82,29 +82,14 @@ export class TableLocalChanges<TData extends { id: string; [key: string]: any }>
             });
         }
         return (
-<<<<<<< HEAD
             <RouterPrompt
                 message={() => {
                     const isDirty = Object.keys(this.state.changes).length > 0 || this.state.changedOrder;
                     if (isDirty) {
-=======
-            <>
-                <RouterPrompt
-                    message={() => {
-                        const isDirty = Object.keys(this.state.changes).length > 0 || this.state.changedOrder;
-                        if (isDirty) {
-                            return "Do you want to save your changes?"; //TODO translate, we need intl context
-                            //return intl.formatMessage(messages.saveUnsavedChanges);
-                        }
-                        return true;
-                    }}
-                    saveAction={async () => {
-                        await this.submitLocalDataChanges();
->>>>>>> main
-                        return true;
+                        return "Do you want to save your changes?"; //TODO translate, we need intl context
+                        //return intl.formatMessage(messages.saveUnsavedChanges);
                     }
-                    return "Do you want to save your changes?"; //TODO translate, we need intl context
-                    //return intl.formatMessage(messages.saveUnsavedChanges);
+                    return true;
                 }}
                 saveAction={async () => {
                     await this.submitLocalDataChanges();