+++ lint-staged.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-    "./*{js,json,md,yml}": "pnpm exec prettier --check",
+    "./!(demo|docs|packages|storybook)/**/*.{js,json,md,yml,yaml}": () => "pnpm lint:root",
 };