+++ packages/admin/cms-admin/src/common/header/Header.tsx
@@ -1,8 +1,6 @@
-import { AppHeader, AppHeaderFillSpace, AppHeaderMenuButton } from "@comet/admin";
+import { AppHeader, AppHeaderFillSpace, AppHeaderMenuButton, CometLogo } from "@comet/admin";
 import { useMediaQuery, useTheme } from "@mui/material";
 import * as React from "react";
-
-import { Logo } from "./Logo";
 
 interface Props {
     children?: React.ReactNode;
@@ -16,7 +14,7 @@ function Header({ children, logo }: Props): React.ReactElement {
     return (
         <AppHeader>
             <AppHeaderMenuButton />
-            {!isMobile && (logo || <Logo />)}
+            {!isMobile && (logo || <CometLogo color="white" />)}
             <AppHeaderFillSpace />
             {children}
         </AppHeader>