+++ packages/api/cms-api/src/dam/dam.module.ts
@@ -8,6 +8,7 @@ import { PixelImageBlockTransformerService } from "./blocks/pixel-image-block-tr
 import { SvgImageBlockTransformerService } from "./blocks/svg-image-block-transformer.service";
 import { DamVideoBlockTransformerService } from "./blocks/video/dam-video-block-transformer.service";
 import { ScaledImagesCacheService } from "./cache/scaled-images-cache.service";
+import { HasValidFilenameConstraint } from "./common/decorators/has-valid-filename.decorator";
 import { DamConfig } from "./dam.config";
 import { DAM_CONFIG, DAM_FILE_VALIDATION_SERVICE, IMGPROXY_CONFIG } from "./dam.constants";
 import { createDamItemsResolver } from "./files/dam-items.resolver";
@@ -130,6 +131,7 @@ export class DamModule {
                 SvgImageBlockTransformerService,
                 DamVideoBlockTransformerService,
                 DamFileDownloadLinkBlockTransformerService,
+                HasValidFilenameConstraint,
             ],
             controllers: [createFilesController({ Scope }), FoldersController, ImagesController],
             exports: [