+++ packages/admin/admin-react-select/CHANGELOG.md
@@ -1,6 +1,5 @@
 # @comet/admin-react-select
 
-<<<<<<< HEAD
 ## 7.0.0-beta.6
 
 ### Patch Changes
@@ -198,7 +197,7 @@
 -   Updated dependencies [b0249e3bc]
 -   Updated dependencies [92eae2ba9]
     -   @comet/admin@7.0.0-beta.0
-=======
+
 ## 6.16.0
 
 ### Patch Changes
@@ -207,7 +206,6 @@
 -   Updated dependencies [747fe32cc]
     -   @comet/admin@6.16.0
     -   @comet/admin-icons@6.16.0
->>>>>>> main
 
 ## 6.15.1