+++ packages/admin/admin-react-select/CHANGELOG.md
@@ -198,6 +198,15 @@
 -   Updated dependencies [92eae2ba9]
     -   @comet/admin@7.0.0-beta.0
 
+## 6.16.0
+
+### Patch Changes
+
+-   Updated dependencies [fb0fe2539]
+-   Updated dependencies [747fe32cc]
+    -   @comet/admin@6.16.0
+    -   @comet/admin-icons@6.16.0
+
 ## 6.15.1
 
 ### Patch Changes